Move a test from Integration to Unit.
